Fix AP disconnect message
This fixes the condition (always indicated when the autopilot comes off), but
it does not fix the sound and warning. We need a separate warning light for
this that will flash master warning, but only temporarily and will not play
the cricket sound, only the cavalry charge.

This fixes the condition for #5, but not the presentation.