Handle invalid waypoint in FMGC.
If the current waypoint is changed (via standard Route Manager menu),
the route manager comes back with current waypoint -1 and trying to
use that as index produces exceptions.

So adjust the conditions so that the code trying to work with waypoints
is not entered in such case. Now the route manager can recover once
the route is corrected again.