Add master warning for nasal exceptions.
When an exception is thrown from any of the nasal loops, something probably
stopped working. So tell the pilot by sounding master caution.

The loop also now does not stop, but is scheduled again. This may lead to
the error repeating, but perhaps when the condition is cleared, it will
return to working.