Default runways when initializing a flight plan.
In FG 3.7 (tested in version/3.7.0-7-gf4fa687) when the flight plan is
initialized without departure runway, the departure waypoint gets
`/autopilot/route-manager/route/wp[0]/leg-bearing-true-deg` = NaN (comes as
`nil` in Nasal) and from that `…/wp[1]/leg-distance-nm` gets NaN and
`/autopilot/route-manager/total-distance` sums as NaN as wel. And that will
cause exceptions in Nasal when it tries to use that value.

Also in both FG3.4 (version/3.4.0-final) and FG3.7 the last leg gets
definite but nonsensical (absurdly large) distance
(`…/wp[n]/leg-distance-nm`) and in the map dialog the segment gets drawn as
loop going to somewhere about equator and back.

So this commit:

 - Defaults both departure and destination runways, choosing from aircraft
   position. For departure it is the correct runway if you start on a runway
   and is the same method as the route manager dialog uses. For the
   destination it is not very good, but the pilot should override the choice
   anyway.

   This shouldn't affect realism much, because to insert SID and STAR via the
   FMGC you have to go over the runway selection dialog anyway. It should
   only give more sensible distance before you do.

 - Moves the setting of departure and destination before inserting the
   waypoints, because otherwise the NaNs would temporarily appear during the
   setup.

Fixes #8.