Avoid exceptions from f_pln when not initialized.
When flight plan is not initialized, f_pln.update_disp was throwing
exceptions, because the f_pln.route_manager member was only set in the
initialization. Since it was only ever set to fmgc.RouteManager, which is
a global anyway, I removed the member and changed all references to go
directly to fmgc.RouteManager.